update backend architecure
