+++ backend/requirements.txt
@@ -10,3 +10,6 @@ waitress
 flask-marshmallow
 flask-sqlalchemy
 marshmallow-sqlalchemy
+Sphinx
+sphinx-rtd-theme
+sphinx-tabs
